feat: [CDS-42766]: Handle ECS sameAsAlreadyRunning, forceDeployment, Empty Scaling Config, Steady State check Logging (#37368)
* feat: [CDS-42766]: Ecs Empty Scalable Target Fix

* feat: [CDS-42766]: Fixed wait for steady state check logging

* feat: [CDS-42766]: Fixed wait for steady state check logging

* feat: [CDS-42766]: Fixed wait for steady state check logging

* feat: [CDS-42766]: Handle same as already running instances, forceNew Deployment

* feat: [CDS-42766]: Handle same as already running instances, forceNew Deployment

* feat: [CDS-42766]: Handle same as already running instances, forceNew Deployment

+++ 960-api-services/src/main/java/io/harness/aws/v2/ecs/EcsV2Client.java
@@ -44,7 +44,7 @@ public interface EcsV2Client {
       AwsInternalConfig awsConfig, RegisterTaskDefinitionRequest registerTaskDefinitionRequest, String region);
 
   WaiterResponse<DescribeServicesResponse> ecsServiceSteadyStateCheck(AwsInternalConfig awsConfig,
-      DescribeServicesRequest describeServicesRequest, String region, int serviceSteadyStateTimeout);
+      DescribeServicesRequest describeServicesRequest, String region, long serviceSteadyStateTimeout);
 
   WaiterResponse<DescribeServicesResponse> ecsServiceInactiveStateCheck(AwsInternalConfig awsConfig,
       DescribeServicesRequest describeServicesRequest, String region, int serviceInactiveStateTimeout);
